usbd: remove compiler warnings
cmd_usbd.c: In function 'write_file_mmc_part':
cmd_usbd.c:644: warning: unused variable 'tmp'

cmd_usbd.c:723: warning: passing argument 2 of 'memsize_parse' from incompatible pointer type
cmd_usbd.c:676: note: expected 'const char **' but argument is of type 'char **'

cmd_usbd.c: In function 'write_mmc_image':
cmd_usbd.c:741: warning: format '%x' expects type 'unsigned int', but argument 3 has type 'long unsigned int'
cmd_usbd.c:742: warning: format '%x' expects type 'unsigned int', but argument 3 has type 'long unsigned int'

cmd_usbd.c:653: note: initialized from here
cmd_usbd.c:1405: warning: 'usbd' may be used uninitialized in this function

+++ include/usbd.h
@@ -125,9 +125,9 @@ struct usbd_ops {
 
 	/* mmc device info */
 	uint mmc_dev;
-	ulong mmc_max;
-	ulong mmc_blk;
-	ulong mmc_total;
+	uint mmc_max;
+	uint mmc_blk;
+	uint mmc_total;
 
 	void (*set_progress)(int);
 	void (*cpu_reset)(void);
